WIP: Fix async window open
